requirements.txt
pandas>=2.0.0
numpy>=1.25.0
stable-baselines3>=2.0.0
gymnasium>=0.28.0
torch>=2.0.0
pyyaml>=6.0
# Use: pip install "C:\Users\mubas\Downloads\ta_lib-0.6.3-cp312-cp312-win_amd64.whl"
# Alternative: talib-binary>=0.4.0